agent-core: align schemas/ + format/ + types/ layout with upstream
Restructure the package to match tldraw/agent-template's directory
shape, motivated by the new "treat the upstream link as live
infrastructure for syncing forward" guidance — file-by-file mapping
makes future syncs easier.

Before:
- schemas/actions.ts        (god-file: visual primitives + frame
                             actions + agent actions all mixed)
- schemas/parts.ts          (parts + AgentPrompt envelope + parser)
- types.ts                  (Streaming<T> + AgentEnv mixed)
- client/convert-shape.ts   (both directions in one file)

After:
- format/                          (visual primitives, mirrors
                                    upstream's shared/format/)
  - focused-color.ts
  - focused-easing.ts              (Anipres-only)
  - focused-frame-action.ts        (Anipres-only)
  - focused-shape.ts
  - convert-tldraw-shape-to-focused-shape.ts
- schemas/                         (agent-facing schemas, mirrors
                                    upstream's shared/schema/)
  - agent-action.ts                (was the agent-action half of
                                    actions.ts)
  - prompt-part.ts                 (renamed from parts.ts)
  - build-response-schema.ts       (unchanged)
- types/                           (mirrors upstream's shared/types/)
  - streaming.ts                   (was Streaming<T> in types.ts)
  - agent-env.ts                   (Anipres-only, was in types.ts)
- client/convert-shape.ts          (slimmed to the Anipres-original
                                    focused→tldraw direction +
                                    slide auto-cue handling)

External barrel surface (agent-core, agent-core/server,
agent-core/client, agent-core/react) is unchanged — downstream
packages keep importing the same names. Only internal deep imports
moved.

Updates THIRD_PARTY_NOTICES.md with the new file paths and a new
entry calling out the format/ vs schemas/ split as itself an
upstream-mirroring choice.

All 30 agent-core tests + 51 worker tests pass; lint, typecheck, and
build clean across agent-core, agent-cli, agent-mcp, worker, app.
