+++ packages/agent-core/src/client/actions/create-shape.ts
@@ -1,4 +1,4 @@
-import type { CreateAction } from "../../schemas/actions.js";
+import type { CreateAction } from "../../schemas/agent-action.js";
 import { registerActionUtil } from "../action-util.js";
 import { focusedShapeToTldrawShape } from "../convert-shape.js";
